Fix SDL event filter and VR texture registration typing

+++ Descent3/vr.cpp
@@ -237,7 +237,7 @@ void VR_EnsureMenuBitmap() {
 
   Vr_menu_bitmap = bm_AllocBitmap(desired_texture_size, desired_texture_size, 0);
   if (Vr_menu_bitmap >= 0) {
-    rend_RegisterExternalTexture(Vr_menu_bitmap, static_cast<uintptr_t>(Vr_menu_fbo_texture), desired_texture_size,
+    rend_RegisterExternalTexture(Vr_menu_bitmap, static_cast<unsigned int>(Vr_menu_fbo_texture), desired_texture_size,
                                  desired_texture_size);
     Vr_menu_texture_registered = true;
   }
